Adding Lerna and moving files around
