+++ package.json
@@ -1,11 +1,10 @@
 {
-  "name": "smcloudstore",
-  "version": "0.0.1",
-  "description": "Multi-cloud object storage module: AWS, Azure, GCP...",
-  "main": "dist/index.js",
+  "name": "smcloudstore-builder",
+  "private": true,
+  "version": "0.1.0",
   "scripts": {
-    "compile": "./node_modules/.bin/tsc",
-    "tsc": "./node_modules/.bin/tsc",
+    "compile": "npm run tsc",
+    "tsc": "sh package.sh",
     "test": "npm run compile && ./node_modules/.bin/_mocha",
     "tslint": "find src -name *.ts -type f -exec './node_modules/.bin/tslint' {} \\;",
     "eslint": "./node_modules/.bin/eslint -c .eslintrc.yml .",
@@ -14,37 +13,20 @@
     "docs": "npm run typedoc",
     "prepublishOnly": "npm run test"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ItalyPaleAle/SMCloudStore.git"
-  },
-  "keywords": [
-    "cloud-storage",
-    "cloud",
-    "object-storage",
-    "s3",
-    "azure",
-    "gcp"
-  ],
   "author": "Alessandro Segala",
   "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/ItalyPaleAle/SMCloudStore/issues"
-  },
   "homepage": "https://github.com/ItalyPaleAle/SMCloudStore",
   "dependencies": {
-    "@google-cloud/storage": "^1.7.0",
-    "azure-storage": "^2.10.1",
-    "minio": "^6.0.0"
   },
   "devDependencies": {
     "@types/google-cloud__storage": "^1.7.0",
     "@types/minio": "^6.0.2",
-    "@types/node": "^10.5.6",
+    "@types/node": "^10.5.7",
     "coveralls": "^3.0.2",
     "digest-stream": "^2.0.0",
     "eslint": "^5.3.0",
     "istanbul": "v1.1.0-alpha.1",
+    "lerna": "^3.0.0-rc.0",
     "mocha": "^5.2.0",
     "mocha-lcov-reporter": "^1.3.0",
     "randomstring": "^1.1.5",
