+++ packages/smcloudstore-azure-storage/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "smcloudstore-azure-storage",
+  "version": "0.1.0",
+  "description": "> TODO: description",
+  "author": "Alessandro Segala <ItalyPaleAle@users.noreply.github.com>",
+  "homepage": "https://github.com/ItalyPaleAle/SMCloudStore/tree/master/packages/smcloudstore-azure-storage",
+  "license": "MIT",
+  "main": "lib/AzureStorageProvider.js",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ItalyPaleAle/SMCloudStore.git"
+  },
+  "scripts": {
+    "test": "echo \"Error: run tests from root\" && exit 1"
+  },
+  "bugs": {
+    "url": "https://github.com/ItalyPaleAle/SMCloudStore/issues"
+  },
+  "dependencies": {
+    "smcloudstore": "^0.1.0",
+    "azure-storage": "^2.10.1"
+  }
+}
