+++ packages/smcloudstore-generic-s3/src/GenericS3Provider.ts
@@ -2,7 +2,7 @@
 
 import {Client as MinioClient, ClientOptions as MinioClientOptions} from 'minio'
 import {Stream} from 'stream'
-import {ListItemObject, ListItemPrefix, ListResults, StorageProvider} from '../lib/StorageProvider'
+import {ListItemObject, ListItemPrefix, ListResults, StorageProvider} from 'smcloudstore/lib/StorageProvider'
 
 /**
  * Client to interact with a generic S3 object storage server, using the Minio library.
