+++ packages/smcloudstore-minio/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "smcloudstore-minio",
+  "version": "0.1.0",
+  "description": "> TODO: description",
+  "author": "Alessandro Segala <ItalyPaleAle@users.noreply.github.com>",
+  "homepage": "https://github.com/ItalyPaleAle/SMCloudStore/tree/master/packages/smcloudstore-minio",
+  "license": "MIT",
+  "main": "lib/MinioProvider.js",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ItalyPaleAle/SMCloudStore.git"
+  },
+  "scripts": {
+    "test": "echo \"Error: run tests from root\" && exit 1"
+  },
+  "bugs": {
+    "url": "https://github.com/ItalyPaleAle/SMCloudStore/issues"
+  },
+  "dependencies": {
+    "smcloudstore-generic-s3": "^0.1.0"
+  }
+}
