+++ packages/smcloudstore/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "smcloudstore",
+  "version": "0.1.0",
+  "description": "Multi-cloud object storage module: AWS, Azure, GCP...",
+  "main": "index.js",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ItalyPaleAle/SMCloudStore.git"
+  },
+  "keywords": [
+    "cloud-storage",
+    "cloud",
+    "object-storage",
+    "s3",
+    "azure",
+    "gcp"
+  ],
+  "author": "Alessandro Segala",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/ItalyPaleAle/SMCloudStore/issues"
+  },
+  "homepage": "https://github.com/ItalyPaleAle/SMCloudStore",
+  "peerDependencies": {
+    "smcloudstore-aws-s3": "^0.1.0",
+    "smcloudstore-azure-storage": "^0.1.0",
+    "smcloudstore-generic-s3": "^0.1.0",
+    "smcloudstore-google-cloud-storage": "^0.1.0",
+    "smcloudstore-minio": "^0.1.0"
+  }
+}
