+++ tsconfig.json
@@ -1,10 +1,10 @@
 {
     "compilerOptions": {
-        "outDir": "./dist",
         "target": "es2016",
         "module": "commonjs",
         "esModuleInterop": true,
         "moduleResolution": "node",
+        "listFiles": true,
         "sourceMap": true,
         "declaration": true,
         "types": [
@@ -12,7 +12,7 @@
         ]
     },
     "include": [
-        "src/**/*"
+        "packages/*/src/*"
     ],
     "exclude": [
         "node_modules"
